requirements.txt
autopep8==1.5
Babel==2.8.0
Django==2.2.7
django-cors-headers==3.2.1
django-filter==2.2.0
django-filters==0.2.1
django-mail-templated==2.6.5
django-phonenumber-field==4.0.0
django-rest-passwordreset==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djoser==2.0.3
phonenumbers==8.11.2
Pillow
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
